pyproject.toml
[tool.poetry]
name = "dags"
version = "1.0.0"
description = "Contains all DAGs"
authors = ["Kartikey Mullick <kartikey@olvin.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
apache-airflow = "2.5.3"

[tool.poetry.group.lint.dependencies]
isort = "^5.10.1"
black = "^22.8.0"
flake8 = "^5.0.4"
mypy = "^0.971"

[tool.isort]
profile = "black"

[tool.mypy]
# 3rd party import
ignore_missing_imports = true
# dynamic typing
disallow_any_unimported = true
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = true
disallow_any_generics = false
disallow_subclassing_any = true
# platform
python_version = "3.10"
# untyped
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
# None and Optional
no_implicit_optional = true
# Warnings
warn_return_any = false
warn_unreachable = true
# Misc
pretty = true